Create `moxie_asset`
