+++ editor/moxie_asset/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "moxie_asset"
+description = "Asset-kind registry, absolute asset source, and the .mat loader for the MotionGfx editor."
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+bevy = { workspace = true, default-features = true, features = [
+    "serialize",
+] }
+ron = { workspace = true }
+serde = { workspace = true }
+
+[lints]
+workspace = true
