+++ editor/moxie_ui/Cargo.toml
@@ -10,6 +10,7 @@ repository.workspace = true
 bevy_fynix = { workspace = true }
 fynix_mock = { workspace = true }
 motiongfx_interp = { workspace = true }
+moxie_asset = { workspace = true }
 bevy = { workspace = true, default-features = true, features = [
     "bevy_feathers",
     "bevy_ui_widgets",
